Limits

+++ src/AdaptiveTimer.h
@@ -126,7 +126,7 @@ namespace QDirStat
         Payload       _payload;
 
         QElapsedTimer _payloadStopwatch;
-        int           _payloadTime      { 0 };
+        qint64        _payloadTime      { 0 };
 
         int           _delayStage       { 0 };
         Delays        _delays;
